// angular-Degree/src/interfaces/Catalog.interface.ts
export interface Catalog {
  communication: string[];
  americanHistoryI: string[];
  americanHistoryII: string[];
  polSciI: string[];
  polSciII: string[];
  arts: string[];
  lpc: string[];
  socialSci: string[];
  tecm: string[];
  mathCore: string[];
  lifeSci: string[];
  req: string[];
  core: string[];
  breadth: string[];
  electives: string[];
  capstone: string[];
}